♻️ refactor: quizAPI 파일 이름 정리 및 api 통합 #64
